+++ appdirs.py
@@ -454,10 +454,7 @@ def _get_win_folder_from_registry(csidl_name):
     registry for this guarantees us the correct answer for all CSIDL_*
     names.
     """
-    if PY3:
-      import winreg as _winreg
-    else:
-      import _winreg
+    import winreg as _winreg
 
     shell_folder_name = {
         "CSIDL_APPDATA": "AppData",