+++ sapling/std/libs/sapX/__init__.py
@@ -1,5 +1,6 @@
 from sapling.std.call_decorator import call_decorator
-from sapling.objects import String, Class
+from sapling.objects import String, Class, Array
+from .parser import parser
 from .lexer import lexer
 
 
@@ -9,3 +10,7 @@ class sapX:
     @call_decorator({'src': {'type': 'string'}}, req_vm=False)
     def _lexer(self, src: String) -> Class:
         return Class.from_py_cls(lexer(src.value), src.line, src.column)
+    
+    @call_decorator({'stream': {'type': 'tokens'}, 'tokens': {'type': 'array'}}, req_vm=False)
+    def _parser(self, stream: Class, tokens: Array) -> Class:
+        return Class.from_py_cls(parser(stream, tokens.to_py_list()), stream.line, stream.column)