+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 ---
 
+v0.1.1
+    - Augmented assignment operators
+    - Indexing
+    - Annotated variables
+    - Bug fixes
 v0.1
     - Enum Support
     - Array Comprehensions